FIX: change the error message and move the error to its module
